test(vrt): deduplicate screenshot helper across suites
