+++ packages/app-nextjs/tests/vrt-helpers.ts
@@ -1,19 +1,9 @@
-import { expect, type Locator, type Page } from '@playwright/test';
+import { type Page } from '@playwright/test';
 
-type VrtScreenshotOptions = {
-	maxDiffPixels?: number;
-};
+export { expectVrtScreenshot } from '../../ontario-design-system-component-library/src/utils/tests/vrt-screenshot';
 
 export const waitForInteractionPaint = async (page: Page) => {
 	await page.evaluate(
 		() => new Promise<void>((resolve) => requestAnimationFrame(() => requestAnimationFrame(() => resolve()))),
 	);
-};
-
-export const expectVrtScreenshot = async (locator: Locator, options: VrtScreenshotOptions = {}) => {
-	await expect(locator).toHaveScreenshot({
-		animations: 'disabled',
-		caret: 'hide',
-		...options,
-	});
 };
