+++ packages/ontario-design-system-component-library/src/utils/tests/vrt-screenshot.ts
@@ -1,0 +1,13 @@
+import { expect, type Locator } from '@playwright/test';
+
+export type VrtScreenshotOptions = {
+	maxDiffPixels?: number;
+};
+
+export const expectVrtScreenshot = async (locator: Locator, options: VrtScreenshotOptions = {}) => {
+	await expect(locator).toHaveScreenshot({
+		animations: 'disabled',
+		caret: 'hide',
+		...options,
+	});
+};
